moved to main component
